클래스 다이어그램 추가, Project 관련 작업(DTO 정의 중)에서 결정사항 : 복합키에 정보를 포함하는 엔티티는 update 없이 모두 삭제 후 요청정보 삽입으로 처리한

+++ src/main/java/me/leewonjun/dewminas/domains/RepositoryLink.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import me.leewonjun.dewminas.domains.compositekeys.RepoLinkPk;
+import me.leewonjun.dewminas.dto.project_sub.RepositoryLinkSummary;
 
 @Getter
 @Setter
@@ -15,12 +16,16 @@ public class RepositoryLink {
     private RepoLinkPk id;
 
     @Column(name = "ropo_type", nullable = false)
-    private Short repoType; // 0 : github
+    private int repoType; // 0 : github
 
     @Builder
-    public RepositoryLink(RepoLinkPk pk, Short repoType) {
+    public RepositoryLink(RepoLinkPk pk, int repoType) {
         this.id = pk;
         this.repoType = repoType;
     }
 
+    // ISP에 근거하여 Updatable을 구현하지 않는다.
+    // 복합키를 가지는 엘리먼트는 모두 삭제 후 삽입할 예정.
+    // Project가 복합키에 포함되어 있어서 리포지토리 전송이 필요함.
+    // 수정 시 기본키 값 수정이 발생 -> 무결성 제약 조건 위배 가능성 있음.
 }
